feat: save report test results only in sqlite database
