+++ .gitignore
@@ -9,7 +9,6 @@ hot
 /lib/static/images
 /lib/static/gui.html
 /lib/static/index.html
-/lib/static/index_sqlite.html
 /lib/static/sql-wasm.js
 /lib/static/sqlite.db
 /lib/static/databaseUrls.json
