+++ lib/common-utils.js
@@ -1,6 +1,5 @@
 'use strict';
 
-const saveFormats = require('./constants/save-formats');
 const {
     SUCCESS,
     FAIL,
@@ -113,8 +112,6 @@ exports.mergeTablesQueries = (dbPaths) => {
 exports.compareDatabaseRowsByTimestamp = (row1, row2) => {
     return row1[exports.dbColumnIndexes.timestamp] - row2[exports.dbColumnIndexes.timestamp];
 };
-
-exports.isSqlite = saveFormat => saveFormat === saveFormats.SQLITE;
 
 exports.formatBrowsers = (collection) => {
     return collection
