+++ lib/config/index.js
@@ -100,7 +100,7 @@ const getParser = () => {
             validate: assertString('path')
         }),
         saveFormat: option({
-            defaultValue: saveFormats.JS,
+            defaultValue: saveFormats.SQLITE,
             validate: assertSaveFormat
         }),
         saveErrorDetails: option({
