+++ lib/constants/save-formats.js
@@ -1,6 +1,5 @@
 'use strict';
 
 module.exports = {
-    JS: 'js',
     SQLITE: 'sqlite'
 };
