+++ lib/gui/tool-runner/index.js
@@ -9,10 +9,8 @@ const Promise = require('bluebird');
 const Runner = require('./runner');
 const subscribeOnToolEvents = require('./report-subscriber');
 const ReportBuilderSqlite = require('../../report-builder/report-builder-sqlite');
-const ReportBuilderJson = require('../../report-builder/report-builder-json');
 const EventSource = require('../event-source');
 const utils = require('../../server-utils');
-const {isSqlite} = require('../../common-utils');
 const {findNode} = require('../../../lib/static/modules/utils');
 const reporterHelper = require('../../reporter-helpers');
 const {UPDATED} = require('../../constants/test-statuses');
@@ -58,19 +56,11 @@ module.exports = class ToolRunner {
     }
 
     async initialize() {
-        this._reuseDataJs = this._readReuseDataJs();
-        const reuseFormat = this._reuseDataJs ? this._reuseDataJs.saveFormat : this._pluginConfig.saveFormat;
-        this.saveResultsToDb = isSqlite(reuseFormat);
+        await mergeDatabasesForReuse(this._reportPath);
 
-        if (this.saveResultsToDb) {
-            await mergeDatabasesForReuse(this._reportPath);
-            this._reportBuilder = ReportBuilderSqlite.create(this._hermione, this._pluginConfig, {reuse: true});
-            await this._reportBuilder.init();
-            await this._reportBuilder.saveStaticFiles();
-        } else {
-            // TODO: remove creating 'report-builder-json' when making saving to DB the default save format
-            this._reportBuilder = ReportBuilderJson.create(this._hermione, this._pluginConfig);
-        }
+        this._reportBuilder = ReportBuilderSqlite.create(this._hermione, this._pluginConfig, {reuse: true});
+        await this._reportBuilder.init();
+        await this._reportBuilder.saveStaticFiles();
 
         this._subscribeOnEvents();
         this._reportBuilder.setApiValues(this._hermione.htmlReporter.values);
@@ -147,24 +137,7 @@ module.exports = class ToolRunner {
         return {suites, browsers};
     }
 
-    _readReuseDataJs() {
-        // TODO: remove loading "data.js" with format save "js"
-        try {
-            return utils.require(path.resolve(this._reportPath, 'data'));
-        } catch (e) {
-            logger.warn(chalk.yellow(`Nothing to reuse in ${this._reportPath}: can not load data.js`));
-            return null;
-        }
-    }
-
     async _loadReuseData() {
-        if (!this._reuseDataJs) {
-            return {};
-        }
-        if (!this.saveResultsToDb) {
-            return this._reuseDataJs;
-        }
-
         const dbPath = path.resolve(this._reportPath, constantFileNames.LOCAL_DATABASE_NAME);
 
         if (await fs.pathExists(dbPath)) {
@@ -172,6 +145,7 @@ module.exports = class ToolRunner {
         }
 
         logger.warn(chalk.yellow(`Nothing to reuse in ${this._reportPath}: can not load data from ${constantFileNames.DATABASE_URLS_JSON_NAME}`));
+
         return {};
     }
