+++ lib/gui/tool-runner/report-subscriber.js
@@ -105,8 +105,6 @@ module.exports = (hermione, reportBuilder, client, reportPath) => {
     });
 
     hermione.on(hermione.events.RUNNER_END, async () => {
-        await reportBuilder.save();
-
         return client.emit(clientEvents.END);
     });
