+++ lib/merge-reports/index.js
@@ -1,24 +1,14 @@
 'use strict';
 
-const {isSqlite} = require('../common-utils');
-const ReportBuilder = require('./report-builder');
 const serverUtils = require('../server-utils');
 
 module.exports = async (pluginConfig, hermione, srcPaths, {destination: destPath}) => {
     validateOpts(srcPaths, destPath);
 
-    if (isSqlite(pluginConfig.saveFormat)) {
-        await Promise.all([
-            serverUtils.saveStaticFilesToReportDir(hermione, pluginConfig, destPath),
-            serverUtils.writeDatabaseUrlsFile(destPath, srcPaths)
-        ]);
-
-        return;
-    }
-
-    const reportBuilder = ReportBuilder.create(srcPaths, destPath);
-
-    await reportBuilder.build();
+    await Promise.all([
+        serverUtils.saveStaticFilesToReportDir(hermione, pluginConfig, destPath),
+        serverUtils.writeDatabaseUrlsFile(destPath, srcPaths)
+    ]);
 };
 
 function validateOpts(srcPaths, destPath) {
