+++ lib/plugin-adapter.js
@@ -3,12 +3,10 @@
 const _ = require('lodash');
 const Promise = require('bluebird');
 const parseConfig = require('./config');
-const ReportBuilderJson = require('./report-builder/report-builder-json');
 const ReportBuilderSqlite = require('./report-builder/report-builder-sqlite');
 const utils = require('./server-utils');
 const cliCommands = require('./cli-commands');
 const PluginApi = require('./plugin-api');
-const {isSqlite} = require('./common-utils');
 
 module.exports = class PluginAdapter {
     static create(hermione, opts) {
@@ -41,44 +39,29 @@ module.exports = class PluginAdapter {
         return this;
     }
 
-    init(prepareData, prepareImages) {
-        this._hermione.on(this._hermione.events.INIT, () => this._run(prepareData, prepareImages));
+    init(prepareData) {
+        this._hermione.on(this._hermione.events.INIT, () => this._run(prepareData));
 
         return this;
     }
 
-    _createReportBuilder(prepareData, prepareImages) {
-        const reportBuilder = ReportBuilderJson.create(this._hermione, this._config);
-
-        return Promise
-            .all([
-                prepareData(this._hermione, reportBuilder, this._config),
-                prepareImages(this._hermione, reportBuilder, this._config)
-            ])
-            .then(() => reportBuilder.save());
-    }
-
-    async _createReportBuilderSqlite(prepareData, prepareImages) {
+    async _createReportBuilderSqlite(prepareData) {
         const reportBuilderSqlite = ReportBuilderSqlite.create(this._hermione, this._config);
+
         await reportBuilderSqlite.init();
 
         return Promise
             .all([
                 reportBuilderSqlite.saveStaticFiles(),
-                prepareData(this._hermione, reportBuilderSqlite, this._config),
-                prepareImages(this._hermione, reportBuilderSqlite, this._config)
+                prepareData(this._hermione, reportBuilderSqlite, this._config)
             ])
             .then(() => reportBuilderSqlite.finalize());
     }
 
-    async _run(prepareData, prepareImages = _.noop) {
-        const generateReport = isSqlite(this._config.saveFormat)
-            ? this._createReportBuilderSqlite(prepareData, prepareImages)
-            : this._createReportBuilder(prepareData, prepareImages);
+    async _run(prepareData) {
+        const generateReport = this._createReportBuilderSqlite(prepareData);
 
-        const endRunnerEvent = this._hermione.events.RUNNER_END || this._hermione.events.END_RUNNER;
-
-        this._hermione.on(endRunnerEvent, () =>
+        this._hermione.on(this._hermione.events.RUNNER_END, () =>
             generateReport.then(() => utils.logPathToHtmlReport(this._config)).catch(utils.logError),
         );
     }
