+++ lib/report-builder/report-builder-sqlite.js
@@ -8,7 +8,6 @@ const {IDLE, RUNNING} = require('../constants/test-statuses');
 const {LOCAL_DATABASE_NAME} = require('../constants/file-names');
 const SqliteAdapter = require('../sqlite-adapter');
 const ReportBuilder = require('./report-builder');
-const saveFormats = require('../constants/save-formats');
 const {
     saveStaticFilesToReportDir,
     writeDatabaseUrlsFile
@@ -75,10 +74,6 @@ module.exports = class ReportBuilderSqlite extends ReportBuilder {
         this._sqliteAdapter.write({testResult, suitePath, suiteName});
     }
 
-    save() {
-        return Promise.resolve(); // added for code compatibility in report-subscriber when working in gui mode
-    }
-
     async finalize() {
         this._sqliteAdapter.close();
 
@@ -91,9 +86,5 @@ module.exports = class ReportBuilderSqlite extends ReportBuilder {
             await writeDatabaseUrlsFile(reportDir, [dbPath]);
             await fs.remove(src);
         }
-    }
-
-    getSaveFormat() {
-        return saveFormats.SQLITE;
     }
 };
