+++ lib/report-builder/report-builder.js
@@ -231,10 +231,6 @@ module.exports = class ReportBuilder {
         return this;
     }
 
-    getSaveFormat() {
-        throw new Error('"getSaveFormat" must be implemented in child class');
-    }
-
     getResult() {
         const {
             defaultView, baseHost, scaleImages, lazyLoadOffset,
@@ -242,6 +238,7 @@ module.exports = class ReportBuilder {
         } = this._pluginConfig;
 
         this._sortTree();
+
         return _.extend({
             skips: _.uniq(this._skips, JSON.stringify),
             suites: this._tree.children,
@@ -251,7 +248,6 @@ module.exports = class ReportBuilder {
             },
             apiValues: this._apiValues,
             date: new Date().toString(),
-            saveFormat: this.getSaveFormat(),
             browsers: this._browsers
         }, this._stats);
     }
