+++ lib/server-utils.js
@@ -8,7 +8,6 @@ const _ = require('lodash');
 const fs = require('fs-extra');
 
 const {ERROR, UPDATED, IDLE, SKIPPED} = require('./constants/test-statuses');
-const saveFormats = require('./constants/save-formats');
 const {IMAGES_PATH} = require('./constants/paths');
 const DATA_FILE_NAME = 'data.js';
 
@@ -66,6 +65,7 @@ function logPathToHtmlReport(pluginConfig) {
     const reportPath = `file://${path.resolve(pluginConfig.path, 'index.html')}`;
 
     logger.log(`Your HTML report is here: ${chalk.yellow(reportPath)}`);
+    logger.log(`To open it use: ${chalk.yellow('npx hermione gui')} or ${chalk.yellow(`npx http-server ${pluginConfig.path}`)}`);
 }
 
 function logError(e) {
@@ -103,7 +103,7 @@ async function saveStaticFilesToReportDir(hermione, pluginConfig, destPath) {
     await Promise.all([
         fs.writeFile(path.resolve(destPath, DATA_FILE_NAME), prepareCommonJSData(getConfig(hermione, pluginConfig)), 'utf8'),
         copyToReportDir(destPath, ['report.min.js', 'report.min.css'], staticFolder),
-        fs.copy(path.resolve(staticFolder, 'index_sqlite.html'), path.resolve(destPath, 'index.html')),
+        fs.copy(path.resolve(staticFolder, 'index.html'), path.resolve(destPath, 'index.html')),
         // as soon as the pull request to sql.js will be accepted, we will switch to using the original package
         fs.copy(require.resolve('@gemini-testing/sql.js'), path.resolve(destPath, 'sql-wasm.js'))
     ]);
@@ -152,8 +152,7 @@ function getConfig(hermione, pluginConfig) {
         skips: [],
         config: {defaultView, baseHost, scaleImages, lazyLoadOffset, errorPatterns, metaInfoBaseUrls, databaseUrlsFile},
         apiValues: hermione.htmlReporter.values,
-        date: new Date().toString(),
-        saveFormat: saveFormats.SQLITE
+        date: new Date().toString()
     };
 }
