+++ lib/static/components/header/summary/index.js
@@ -6,7 +6,6 @@ import PropTypes from 'prop-types';
 import SummaryKey from './item';
 import DbSummaryKey from './dbSummary';
 import {getStats} from '../../../modules/utils';
-import {isSqlite} from '../../../../common-utils';
 
 import './summary.css';
 
@@ -23,7 +22,7 @@ class Summary extends Component {
 
     render() {
         const {total, passed, failed, skipped, retries} = this.props.stats;
-        const {useSqlite, fetchDbDetails} = this.props;
+        const {fetchDbDetails} = this.props;
 
         return (
             <dl className="summary">
@@ -32,7 +31,7 @@ class Summary extends Component {
                 <SummaryKey label="Failed" id="failed" value={failed} />
                 <SummaryKey label="Retries" id="retries" value={retries} />
                 <SummaryKey label="Skipped" id="skipped" value={skipped} />
-                {useSqlite ? <DbSummaryKey fetchDbDetails={fetchDbDetails}/> : null}
+                <DbSummaryKey fetchDbDetails={fetchDbDetails} />
             </dl>
         );
     }
@@ -46,7 +45,6 @@ export default connect(
 
         return {
             stats: statsToShow,
-            useSqlite: isSqlite(state.saveFormat),
             fetchDbDetails: state.fetchDbDetails
         };
     })(Summary);
