+++ lib/static/components/main-tree.js
@@ -9,7 +9,6 @@ import ErrorGroupsList from './error-groups/list';
 import Suites from './suites';
 import clientEvents from '../../gui/constants/client-events';
 import {openDbConnection, closeDbConnection, suiteBegin, testBegin, testResult, testsEnd} from '../modules/actions';
-import {isSqlite} from '../../common-utils';
 import Loading from './loading';
 
 class MainTree extends Component {
@@ -20,13 +19,14 @@ class MainTree extends Component {
 
     componentDidMount() {
         this.props.gui && this._subscribeToEvents();
-        if (this.props.useSqlite) {
+
+        if (!this.props.gui) {
             this.props.actions.openDbConnection();
         }
     }
 
     componentWillUnmount() {
-        if (this.props.useSqlite) {
+        if (!this.props.gui) {
             this.props.actions.closeDbConnection();
         }
     }
@@ -57,9 +57,9 @@ class MainTree extends Component {
     }
 
     render() {
-        const {groupByError, suites, useSqlite, fetchDbDetails} = this.props;
+        const {groupByError, suites, fetchDbDetails} = this.props;
 
-        if (!Object.keys(suites).length && !fetchDbDetails && useSqlite) {
+        if (!Object.keys(suites).length && !fetchDbDetails) {
             return (<Loading active={true}/>);
         }
 
@@ -77,7 +77,6 @@ export default connect(
         return ({
             gui: state.gui,
             groupByError,
-            useSqlite: isSqlite(state.saveFormat),
             fetchDbDetails: state.fetchDbDetails,
             suites: state.suites
         });
