+++ lib/static/modules/reducers/reporter.js
@@ -32,7 +32,7 @@ const compiledData = window.data || defaultState;
 function getInitialState(data) {
     const {
         skips, suites, config, total, updated, passed, browsers,
-        failed, skipped, warned, retries, perBrowser, apiValues, gui = false, autoRun, date, saveFormat
+        failed, skipped, warned, retries, perBrowser, apiValues, gui = false, autoRun, date
     } = data;
 
     config.errorPatterns = config.errorPatterns.map((patternInfo) => ({...patternInfo, regexp: new RegExp(patternInfo.pattern)}));
@@ -45,7 +45,7 @@ function getInitialState(data) {
     }
 
     let formattedSuites = {};
-    // if saveFormat == sqlite then suites will be defined only at gui mode
+
     if (suites) {
         formattedSuites = formatSuitesData(suites);
     }
@@ -70,7 +70,6 @@ function getInitialState(data) {
         config,
         apiValues,
         date: dateToLocaleString(date),
-        saveFormat,
         stats: {
             all: {total, updated, passed, failed, skipped, retries, warned},
             perBrowser
