+++ test/func/fixtures/failed-describe.hermione.js
