+++ test/func/fixtures/success-describe.hermione.js
