+++ test/unit/lib/merge-reports/index.js
@@ -1,23 +1,22 @@
 'use strict';
 
 const _ = require('lodash');
-const Promise = require('bluebird');
 const {stubTool, stubConfig} = require('../../utils');
 const mergeReports = require('lib/merge-reports');
-const ReportBuilder = require('lib/merge-reports/report-builder');
+const serverUtils = require('lib/server-utils');
 
 describe('lib/merge-reports', () => {
     const sandbox = sinon.sandbox.create();
 
-    const execMergeReports_ = async (paths = [], opts = {}) => {
+    const execMergeReports_ = async ({pluginConfig = stubConfig(), hermione = stubTool(stubConfig()), paths = [], opts = {}}) => {
         opts = _.defaults(opts, {destination: 'default-dest-report/path'});
 
-        await mergeReports(stubConfig(), stubTool(stubConfig()), paths, opts);
+        await mergeReports(pluginConfig, hermione, paths, opts);
     };
 
     beforeEach(() => {
-        sandbox.stub(ReportBuilder, 'create').returns(Object.create(ReportBuilder.prototype));
-        sandbox.stub(ReportBuilder.prototype, 'build').resolves();
+        sandbox.stub(serverUtils, 'saveStaticFilesToReportDir').resolves();
+        sandbox.stub(serverUtils, 'writeDatabaseUrlsFile').resolves();
     });
 
     afterEach(() => sandbox.restore());
@@ -25,35 +24,28 @@ describe('lib/merge-reports', () => {
     describe('options validation', () => {
         it('should throw error if no source reports paths are specified', async () => {
             await assert.isRejected(
-                execMergeReports_([]),
+                execMergeReports_({paths: []}),
                 'Nothing to merge, no source reports are passed'
             );
         });
 
         it('should throw error if destination report path exists in passed source reports paths', async () => {
             await assert.isRejected(
-                execMergeReports_(['src-report/path', 'dest-report/path'], {destination: 'dest-report/path'}),
+                execMergeReports_({paths: ['src-report/path', 'dest-report/path'], opts: {destination: 'dest-report/path'}}),
                 'Destination report path: dest-report/path, exists in source report paths'
             );
         });
     });
 
-    it('should not fail if only one source report is specified', async () => {
-        await assert.isFulfilled(execMergeReports_(['src-report/path']));
-    });
+    it('should merge reports', async () => {
+        const pluginConfig = stubConfig();
+        const hermione = stubTool(pluginConfig);
+        const paths = ['src-report/path-1', 'src-report/path-2'];
+        const destination = 'dest-report/path';
 
-    it('should create "ReportBuilder" instance with passed sources and destination paths', async () => {
-        await execMergeReports_(['src-report/path-1', 'src-report/path-2'], {destination: 'dest-report/path'});
+        await execMergeReports_({pluginConfig, hermione, paths, opts: {destination}});
 
-        assert.calledOnceWith(ReportBuilder.create, ['src-report/path-1', 'src-report/path-2'], 'dest-report/path');
-    });
-
-    it('should wait until build report will be finished', async () => {
-        const afterBuild = sinon.spy().named('afterBuild');
-        ReportBuilder.prototype.build.callsFake(() => Promise.delay(10).then(afterBuild));
-
-        await execMergeReports_(['src-report/path-1', 'src-report/path-2'], {destination: 'dest-report/path'});
-
-        assert.calledOnce(afterBuild);
+        assert.calledOnceWith(serverUtils.saveStaticFilesToReportDir, hermione, pluginConfig, destination);
+        assert.calledOnceWith(serverUtils.writeDatabaseUrlsFile, destination, paths);
     });
 });
